requirements.txt
streamlit
ultralytics
python-dotenv
easyocr
torch
torchvision
torchaudio
Pillow
numpy
opencv-python
faiss-cpu
clip @ git+https://github.com/openai/CLIP.git
